fix merge conflicts (#2115)
