+++ .github/workflows/protocol-build-and-push.yml
@@ -3,47 +3,12 @@ name: Protocol Build & Push Image to AWS ECR
 on:  # yamllint disable-line rule:truthy
   push:
     branches:
+      - 'wl/5.2.x-fns'
       - main
       - 'release/protocol/v[0-9]+.[0-9]+.x'  # e.g. release/protocol/v0.1.x
       - 'release/protocol/v[0-9]+.x'  # e.g. release/protocol/v1.x
 
 jobs:
-  build-and-push-dev:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./protocol
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: '0'  # without this, ignite fails.
-
-      - name: Configure AWS credentials
-        uses: aws-actions/configure-aws-credentials@v1
-        with:
-          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID_VALIDATOR_DEV }}
-          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY_VALIDATOR_DEV }}
-          aws-region: us-east-2
-
-      - name: Login to Amazon ECR
-        id: login-ecr
-        uses: aws-actions/amazon-ecr-login@v1
-
-      - name: Build, Tag, and Push the Image to Amazon ECR
-        id: build-image
-        env:
-          ECR_REGISTRY: ${{ steps.login-ecr.outputs.registry }}
-          ECR_REPOSITORY: dev-validator
-        run: |
-          make localnet-build-amd64
-          commit_hash=$(git rev-parse --short=7 HEAD)
-          docker build \
-            --platform amd64 \
-            -t $ECR_REGISTRY/$ECR_REPOSITORY:$commit_hash \
-            -f testing/testnet-dev/Dockerfile .
-          docker push $ECR_REGISTRY/$ECR_REPOSITORY --all-tags
-
   build-and-push-dev2:
     runs-on: ubuntu-latest
     defaults:
@@ -78,148 +43,4 @@ jobs:
             --platform amd64 \
             -t $ECR_REGISTRY/$ECR_REPOSITORY:$commit_hash \
             -f testing/testnet-dev/Dockerfile .
-          docker push $ECR_REGISTRY/$ECR_REPOSITORY --all-tags
-
-  build-and-push-dev3:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./protocol
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: '0'  # without this, ignite fails.
-
-      - name: Configure AWS credentials
-        uses: aws-actions/configure-aws-credentials@v1
-        with:
-          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID_VALIDATOR_DEV3 }}
-          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY_VALIDATOR_DEV3 }}
-          aws-region: us-east-2
-
-      - name: Login to Amazon ECR
-        id: login-ecr
-        uses: aws-actions/amazon-ecr-login@v1
-
-      - name: Build, Tag, and Push the Image to Amazon ECR
-        id: build-image
-        env:
-          ECR_REGISTRY: ${{ steps.login-ecr.outputs.registry }}
-          ECR_REPOSITORY: dev3-validator
-        run: |
-          make localnet-build-amd64
-          commit_hash=$(git rev-parse --short=7 HEAD)
-          docker build \
-            --platform amd64 \
-            -t $ECR_REGISTRY/$ECR_REPOSITORY:$commit_hash \
-            -f testing/testnet-dev/Dockerfile .
-          docker push $ECR_REGISTRY/$ECR_REPOSITORY --all-tags
-
-  build-and-push-dev4:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./protocol
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: '0'  # without this, ignite fails.
-
-      - name: Configure AWS credentials
-        uses: aws-actions/configure-aws-credentials@v1
-        with:
-          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID_VALIDATOR_DEV4 }}
-          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY_VALIDATOR_DEV4 }}
-          aws-region: us-east-2
-
-      - name: Login to Amazon ECR
-        id: login-ecr
-        uses: aws-actions/amazon-ecr-login@v1
-
-      - name: Build, Tag, and Push the Image to Amazon ECR
-        id: build-image
-        env:
-          ECR_REGISTRY: ${{ steps.login-ecr.outputs.registry }}
-          ECR_REPOSITORY: dev4-validator
-        run: |
-          make localnet-build-amd64
-          commit_hash=$(git rev-parse --short=7 HEAD)
-          docker build \
-            --platform amd64 \
-            -t $ECR_REGISTRY/$ECR_REPOSITORY:$commit_hash \
-            -f testing/testnet-dev/Dockerfile .
-          docker push $ECR_REGISTRY/$ECR_REPOSITORY --all-tags
-
-  build-and-push-dev5:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./protocol
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: '0'  # without this, ignite fails.
-
-      - name: Configure AWS credentials
-        uses: aws-actions/configure-aws-credentials@v1
-        with:
-          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID_VALIDATOR_DEV5 }}
-          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY_VALIDATOR_DEV5 }}
-          aws-region: us-east-2
-
-      - name: Login to Amazon ECR
-        id: login-ecr
-        uses: aws-actions/amazon-ecr-login@v1
-
-      - name: Build, Tag, and Push the Image to Amazon ECR
-        id: build-image
-        env:
-          ECR_REGISTRY: ${{ steps.login-ecr.outputs.registry }}
-          ECR_REPOSITORY: dev5-validator
-        run: |
-          make localnet-build-amd64
-          commit_hash=$(git rev-parse --short=7 HEAD)
-          docker build \
-            --platform amd64 \
-            -t $ECR_REGISTRY/$ECR_REPOSITORY:$commit_hash \
-            -f testing/testnet-dev/Dockerfile .
-          docker push $ECR_REGISTRY/$ECR_REPOSITORY --all-tags
-
-  build-and-push-staging:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./protocol
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: '0'  # without this, ignite fails.
-
-      - name: Configure AWS credentials
-        uses: aws-actions/configure-aws-credentials@v1
-        with:
-          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID_VALIDATOR_STAGING }}
-          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY_VALIDATOR_STAGING }}
-          aws-region: us-east-2
-
-      - name: Login to Amazon ECR
-        id: login-ecr
-        uses: aws-actions/amazon-ecr-login@v1
-
-      - name: Build, Tag, and Push the Image to Amazon ECR
-        id: build-image
-        env:
-          ECR_REGISTRY: ${{ steps.login-ecr.outputs.registry }}
-          ECR_REPOSITORY: staging-validator
-        run: |
-          make localnet-build-amd64
-          commit_hash=$(git rev-parse --short=7 HEAD)
-          docker build \
-            --platform amd64 \
-            -t $ECR_REGISTRY/$ECR_REPOSITORY:$commit_hash \
-            -f testing/testnet-staging/Dockerfile .
           docker push $ECR_REGISTRY/$ECR_REPOSITORY --all-tags
