+++ protocol/x/clob/types/expected_keepers.go
@@ -81,12 +81,6 @@ type SubaccountsKeeper interface {
 		ctx sdk.Context,
 		perpetualId uint32,
 	) (sdk.AccAddress, error)
-	DistributeFees(
-		ctx sdk.Context,
-		assetId uint32,
-		quantums *big.Int,
-		perpetualId uint32,
-	) error
 	SendSubaccountUpdates(
 		ctx sdk.Context,
 		subaccountUpdates []satypes.StreamSubaccountUpdate,
