+++ protocol/x/subaccounts/keeper/isolated_subaccount_test.go
@@ -16,7 +16,7 @@ import (
 func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 	tests := map[string]struct {
 		// parameters
-		settledUpdateWithUpdatedSubaccount keeper.SettledUpdate
+		settledUpdateWithUpdatedSubaccount types.SettledUpdate
 		perpetuals                         []perptypes.Perpetual
 
 		// expectation
@@ -24,7 +24,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 		expectedErr             error
 	}{
 		`If no perpetual updates, nil state transition is returned`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id:                 &constants.Alice_Num0,
 					PerpetualPositions: nil,
@@ -37,7 +37,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 			expectedStateTransition: nil,
 		},
 		`If single non-isolated perpetual updates, nil state transition is returned`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id:                 &constants.Alice_Num0,
 					PerpetualPositions: nil,
@@ -57,7 +57,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 			expectedStateTransition: nil,
 		},
 		`If multiple non-isolated perpetual updates, nil state transition is returned`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id:                 &constants.Alice_Num0,
 					PerpetualPositions: nil,
@@ -82,7 +82,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 			expectedStateTransition: nil,
 		},
 		`If multiple non-isolated perpetual positions, nil state transition is returned`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id: &constants.Alice_Num0,
 					PerpetualPositions: []*types.PerpetualPosition{
@@ -106,7 +106,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 			expectedStateTransition: nil,
 		},
 		`If single isolated perpetual update, no perpetual position, state transition is returned for closed position`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id:                 &constants.Alice_Num0,
 					PerpetualPositions: nil,
@@ -139,7 +139,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 		},
 		`If single isolated perpetual update, existing perpetual position with same size, state transition is returned for
 		opened position`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id: &constants.Alice_Num0,
 					PerpetualPositions: []*types.PerpetualPosition{
@@ -177,7 +177,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 		},
 		`If single isolated perpetual update, existing perpetual position with different size, nil state transition
 		returned`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id: &constants.Alice_Num0,
 					PerpetualPositions: []*types.PerpetualPosition{
@@ -209,7 +209,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 			expectedStateTransition: nil,
 		},
 		`Returns error if perpetual position was opened with no asset updates`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id: &constants.Alice_Num0,
 					PerpetualPositions: []*types.PerpetualPosition{
@@ -237,7 +237,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 			expectedErr:             types.ErrFailedToUpdateSubaccounts,
 		},
 		`Returns error if perpetual position was opened with multiple asset updates`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id: &constants.Alice_Num0,
 					PerpetualPositions: []*types.PerpetualPosition{
@@ -278,7 +278,7 @@ func TestGetIsolatedPerpetualStateTransition(t *testing.T) {
 			expectedErr:             types.ErrFailedToUpdateSubaccounts,
 		},
 		`Returns error if perpetual position was opened with non-usdc asset update`: {
-			settledUpdateWithUpdatedSubaccount: keeper.SettledUpdate{
+			settledUpdateWithUpdatedSubaccount: types.SettledUpdate{
 				SettledSubaccount: types.Subaccount{
 					Id: &constants.Alice_Num0,
 					PerpetualPositions: []*types.PerpetualPosition{
